Try using requests module to get a website
Use the requests module to access bing and make a search using a
parameter. Something interesting that I learned is that if you name a
python file with the name of a module, it generates a conflict and your
file has priority, so you won't be able to access the module's file.